commitment: route seq-vs-batch tests through processSeq/processBatch, promote processFreshTrie
